==> Notebooks/1. Análisis Exploratorio.ipynb <==
import pandas as pd
import numpy as np
import datetime

pd.options.display.max_columns = None

df = pd.read_excel("../Data/2021_Accidentalidad.xlsx")
df.head(3)

# Exploración inicial de nuestros datos

# Tipado de datos

df.columns

# Tamaño

df.shape

# Tipo de Datos

df.dtypes

# Hagamos una exploración de las variables

df.info()

# Principales estadísticos

df.describe().T

df.describe(include = "object").T

# Duplicados

df.duplicated().sum()

Eliminamos duplicados

df.drop_duplicates(inplace = True)

# Nulos

df.isnull().sum()

Más adelante trabajaremos en las transformaciones de los valores nulos de las columans que los presentan

(df.isnull().sum() / df.shape[0]) * 100
