==> Notebooks/3. Selenium.ipynb <==
# Selenium

import requests
import pandas as pd
from time import sleep
import numpy as np


from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options




import warnings
warnings.filterwarnings('ignore')

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
#para ocultarme como robot
opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized') #empezar maximizado
opciones.add_argument('user.data-dir=selenium') #guarda las cookies
opciones.add_argument('--incognito')#incognito window

# iniciamos el driver
driver = webdriver.Chrome(ChromeDriverManager().install())
## accedemos a la pagina web

driver.get("https://www.wunderground.com/history")
driver.implicitly_wait(10)
# esperamos
driver.find_element_by_css_selector("#truste-consent-button").click()
# aceptamos las cookies

sleep(5)
# seleccionamos la casilla para poner la ciudad por la que queremos buscar
driver.find_element_by_css_selector("#historySearch").send_keys("Madrid, Madrid, Spain", Keys.TAB)
sleep(5)
# por lo que sea le tenemos que dar dos veces al botón de enviar
driver.find_element_by_xpath('//*[@id="dateSubmit"]').click()
sleep(2)
driver.find_element_by_xpath('//*[@id="dateSubmit"]').click()
sleep(5)

driver.find_element_by_xpath('//*[@id="inner-content"]/div[2]/div[1]/div[1]/div[1]/div/lib-link-selector/div/div/div/a[3]').click()
# seleccionamos que nos los valores medios del mes
resultado = driver.find_element_by_css_selector("#inner-content > div.region-content-main > div.row > div:nth-child(5) > div:nth-child(1) > div > lib-city-history-observation > div > div.observation-table.ng-star-inserted").text



url_list_1 = []

for month in range(1, 13):
    url_list_1.append(f"https://www.wunderground.com/history/monthly/LEMD/date/2021-{month}")

result_list_1 = []

for i in url_list_1: 
    driver.get(i)
    try:
        sleep(5)
        driver.find_element_by_css_selector('#truste-consent-button').click()
        sleep(2)
        resultado = driver.find_element_by_css_selector("#inner-content > div.region-content-main > div.row > div:nth-child(5) > div:nth-child(1) > div > lib-city-history-observation > div > div.observation-table.ng-star-inserted").text
        result_list_1.append(resultado)
    except: 
        sleep(5)
        resultado = driver.find_element_by_css_selector("#inner-content > div.region-content-main > div.row > div:nth-child(5) > div:nth-child(1) > div > lib-city-history-observation > div > div.observation-table.ng-star-inserted").text
        result_list_1.append(resultado)
driver.quit()

df_selenium = pd.DataFrame(result_list_1)
df_selenium

df_selenium.columns = ['Prueba']
df_selenium.head(1)

df_selenium.iloc[0][0]

x = df_selenium.iloc[0][0].split("\n")

x

pd.DataFrame(np.array(x[1:]).reshape(7, 32)).T

x[1:].index("Max Avg Min")


x[1]

# hacemos un dataframe gordote
df_clima = pd.DataFrame()

for i in range(len(df_selenium)):
    x = list(df_selenium.iloc[i][0].split("\n"))
    
    mi_ind = x[1:].index("Max Avg Min")
    print(mi_ind)
    
    df_solo =  pd.DataFrame(np.array(x[1:]).reshape(7, mi_ind)).T
    
    df_solo.columns = ["Time", "Temperature (° F)", "Dew Point (° F)", "Humidity (%)", "Wind Speed (mph)", "Pressure (Hg)", "Precipitation (in)"]
    df_solo["month"] = x[1]
    df_clima = pd.concat([df_clima, df_solo], axis = 0)

df_clima.head(400)

df_clima[["Tmax", "Avg", "Tmin"]] = df_clima["Temperature (° F)"].str.split(" ", expand = True)

def separar_columnas(lista, columna):
    df_clima[lista] = df_clima[columna].str.split(" ", expand = True)
    return df_clima[columna].str.split(" ", expand = True)


columnas = df_clima.columns[1:6]
columnas

listas = [["Tmax", "Avg", "Tmin"], ["Dmax", "DAvg", "Dmin"], ["Hmax", "HAvg", "Hmin"], ["Wmax", "WAvg", "Wmin"], 
         ["Pmax", "PAvg", "Pmin"]]

for c, l in zip(listas, columnas):
    print(separar_columnas(c,l))

df_clima.head(2)

df_clima.drop([0], axis = 0, inplace = True)

df_clima.head(2)

df_clima.columns

df_clima.columns = ['dia', 'Temperature (° F)', 'Dew Point (° F)', 'Humidity (%)',
       'Wind Speed (mph)', 'Pressure (Hg)', 'precipitacion(in)', 'month',
       'temperatura(°C)_max', 'temperatura(°C)_avg', 'temperatura(°C)_min', 'pundo_de_condensacion(°C)_max', 'pundo_de_condensacion(°C)_avg', 'pundo_de_condensacion(°C)_min', 'humedad(%)_max', 'humedad(%)_avg', 'humedad(%)_min',
       'vel_viento_max', 'vel_viento_avg', 'vel_viento_min', 'presion(Hg)_max', 'presion(Hg)_avg', 'presion(Hg)_min']
df_clima.head(1)

eliminar = ['Temperature (° F)', 'Dew Point (° F)', 'Humidity (%)',
       'Wind Speed (mph)', 'Pressure (Hg)']

df_clima = df_clima.drop(eliminar, axis = 1)

df_clima.head()

df_clima.dtypes

df_clima["mes"] = df_clima['month'].map({'Jan':1,
                             'Feb':2,
                             'Mar':3,
                            'Apr':4,
                            'May':5,
                            'Jun':6,
                             'Jul':7,
                             'Aug':8,
                            'Sep':9,
                             'Oct':10,
                             'Nov':11,
                            'Dec':12},
                             na_action=None)

year = pd.datetime.now().year
df_clima['fecha']=pd.to_datetime({'year': 2021,
                           'month': df_clima["mes"],
                           'day': df_clima['dia']
                           })

df_clima['fecha'] = pd.to_datetime(df_clima['fecha'])

df_clima.head()

df_clima['temperatura(°C)_max'] = pd.to_numeric(df_clima["temperatura(°C)_max"], errors='coerce').astype('float64')
df_clima['temperatura(°C)_avg'] = pd.to_numeric(df_clima["temperatura(°C)_avg"], errors='coerce').astype('float64')
df_clima['temperatura(°C)_min'] = pd.to_numeric(df_clima["temperatura(°C)_max"], errors='coerce').astype('float64')
df_clima['pundo_de_condensacion(°C)_max'] = pd.to_numeric(df_clima["pundo_de_condensacion(°C)_max"], errors='coerce').astype('float64')
df_clima['pundo_de_condensacion(°C)_avg'] = pd.to_numeric(df_clima["pundo_de_condensacion(°C)_avg"], errors='coerce').astype('float64')
df_clima['pundo_de_condensacion(°C)_min'] = pd.to_numeric(df_clima["pundo_de_condensacion(°C)_min"], errors='coerce').astype('float64')
df_clima['humedad(%)_max'] = pd.to_numeric(df_clima["humedad(%)_max"], errors='coerce').astype('float64')
df_clima['humedad(%)_avg'] = pd.to_numeric(df_clima["humedad(%)_avg"], errors='coerce').astype('float64')
df_clima['humedad(%)_min'] = pd.to_numeric(df_clima["humedad(%)_min"], errors='coerce').astype('float64')
df_clima['presion(Hg)_max'] = pd.to_numeric(df_clima["presion(Hg)_max"], errors='coerce').astype('float64')
df_clima['presion(Hg)_avg'] = pd.to_numeric(df_clima["presion(Hg)_avg"], errors='coerce').astype('float64')
df_clima['presion(Hg)_min'] = pd.to_numeric(df_clima["presion(Hg)_min"], errors='coerce').astype('float64')
df_clima['vel_viento_max'] = pd.to_numeric(df_clima["vel_viento_max"], errors='coerce').astype('float64')
df_clima['vel_viento_avg'] = pd.to_numeric(df_clima["vel_viento_avg"], errors='coerce').astype('float64')
df_clima['vel_viento_min'] = pd.to_numeric(df_clima["vel_viento_min"], errors='coerce').astype('float64')


df_clima.dtypes

df_clima['temperatura(°C)_max'] = round(((df_clima['temperatura(°C)_max'] - 32) *5/9),1)
df_clima['temperatura(°C)_avg'] = round(((df_clima['temperatura(°C)_avg'] - 32) *5/9),1)
df_clima['temperatura(°C)_min'] = round(((df_clima['temperatura(°C)_min'] - 32) *5/9),1)
df_clima['pundo_de_condensacion(°C)_max'] = round(((df_clima['pundo_de_condensacion(°C)_max'] - 32) *5/9),1)
df_clima['pundo_de_condensacion(°C)_avg'] = round(((df_clima['pundo_de_condensacion(°C)_avg'] - 32) *5/9),1)
df_clima['pundo_de_condensacion(°C)_min'] = round(((df_clima['pundo_de_condensacion(°C)_min'] - 32) *5/9),1)



df_clima.head()

# Importamos nuestro CSV limpio y lo enriquecemos con Selenium

df = pd.read_excel("../Data/df_limpio_transformado.xlsx")
df.head(2)

mergeado_inner = df.merge(df_clima, on='fecha')
mergeado_inner.head()

mergeado_inner.columns

eliminar1 = ['Unnamed: 0', 'coordenada_x_utm', 'coordenada_y_utm',
       'precipitacion(in)', 'localizacion', 'numero', 'cod_distrito', 'dia']

mergeado_inner = mergeado_inner.drop(eliminar1, axis = 1)

mergeado_inner.to_excel("df_mergeado.xlsx")

# Importamos CSV con datos de los Distritos

df_d = pd.read_excel("distritos.xlsx")
df_d.head(2)

mergeado_inner1 = mergeado_inner.merge(df_d, on='distrito')
mergeado_inner1.head()

mergeado_inner1.to_excel("../Data/df_mergeado1.xlsx")
